+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.2 (unreleased)
 - better handling of disconnection on modbus socket
+- add crash handler for core dumps
 
 ## 1.1.1 (21.11.2019)
 - reset registers to 0xFFFF when empty